+++ src/app/providers/index.ts
@@ -1,3 +1,4 @@
 export * from './feedback.provider';
 export * from './list-files.providers';
 export * from './list-methods.providers';
+export * from './list-modules.providers';